# www/doc/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "0.0.1",
    "title" : "Petstore Application"
  },
  "host" : "racoon.iot1.etsit.upv.es:8888",
  "basePath" : "/petstore/api",
  "tags" : [ {
    "name" : "Authorization API"
  }, {
    "name" : "Pet Resource"
  } ],
  "schemes" : [ "http" ],
  "paths" : {
    "/authorization/token/generate" : {
      "get" : {
        "tags" : [ "Authorization API" ],
        "summary" : "Generate new authorization access and refresh token",
        "description" : "",
        "operationId" : "generateToken",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/GeneratedToken"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          }
        },
        "security" : [ {
          "basic" : [ ]
        } ]
      }
    },
    "/authorization/token/refresh" : {
      "get" : {
        "tags" : [ "Authorization API" ],
        "summary" : "Generate new access token given the refresh token",
        "description" : "",
        "operationId" : "refreshToken",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/GeneratedToken"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          }
        },
        "security" : [ {
          "refresh" : [ ]
        } ]
      }
    },
    "/authorization/token/revoke/{token}" : {
      "get" : {
        "tags" : [ "Authorization API" ],
        "summary" : "Revoke specified refresh token",
        "description" : "",
        "operationId" : "revokeToken",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "token",
          "in" : "path",
          "description" : "Refresh Token",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "401" : {
            "description" : "Unauthorized"
          }
        },
        "security" : [ {
          "access" : [ ]
        } ]
      }
    },
    "/authorization/token/revokeAll" : {
      "get" : {
        "tags" : [ "Authorization API" ],
        "summary" : "Revoke all refresh tokens",
        "description" : "",
        "operationId" : "revokeAll",
        "produces" : [ "application/json" ],
        "responses" : {
          "401" : {
            "description" : "Unauthorized"
          }
        },
        "security" : [ {
          "access" : [ ]
        } ]
      }
    },
    "/resource/pet/create" : {
      "put" : {
        "tags" : [ "Pet Resource" ],
        "summary" : "Create a pet",
        "description" : "If id is not provided, it will be randomly generated",
        "operationId" : "create",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "Pet description",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Pet"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Pet"
            }
          },
          "400" : {
            "description" : "Pet already exists"
          },
          "500" : {
            "description" : "Internal error"
          }
        },
        "security" : [ {
          "access" : [ ]
        } ]
      }
    },
    "/resource/pet/delete/{id}" : {
      "delete" : {
        "tags" : [ "Pet Resource" ],
        "summary" : "Delete a pet",
        "description" : "",
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID of pet",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "400" : {
            "description" : "Pet not found"
          },
          "500" : {
            "description" : "Internal error"
          }
        },
        "security" : [ {
          "access" : [ ]
        } ]
      }
    },
    "/resource/pet/get/{id}" : {
      "get" : {
        "tags" : [ "Pet Resource" ],
        "summary" : "Get a pet by id",
        "description" : "",
        "operationId" : "get",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID of pet",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Pet"
            }
          },
          "400" : {
            "description" : "Pet not found"
          },
          "500" : {
            "description" : "Internal error"
          }
        },
        "security" : [ {
          "access" : [ ]
        } ]
      }
    },
    "/resource/pet/list" : {
      "get" : {
        "tags" : [ "Pet Resource" ],
        "summary" : "List all pets",
        "description" : "",
        "operationId" : "list",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Pet"
              }
            }
          },
          "500" : {
            "description" : "Internal error"
          }
        },
        "security" : [ {
          "access" : [ ]
        } ]
      }
    },
    "/resource/pet/update" : {
      "post" : {
        "tags" : [ "Pet Resource" ],
        "summary" : "Update a pet",
        "description" : "",
        "operationId" : "update",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "Pet description",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Pet"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Pet"
            }
          },
          "400" : {
            "description" : "Pet not found"
          },
          "500" : {
            "description" : "Internal error"
          }
        },
        "security" : [ {
          "access" : [ ]
        } ]
      }
    }
  },
  "securityDefinitions" : {
    "access" : {
      "type" : "apiKey",
      "name" : "Authorization",
      "in" : "header"
    },
    "basic" : {
      "type" : "basic"
    },
    "refresh" : {
      "type" : "apiKey",
      "name" : "Authorization",
      "in" : "header"
    }
  },
  "definitions" : {
    "GeneratedToken" : {
      "type" : "object",
      "properties" : {
        "accessToken" : {
          "type" : "string"
        },
        "refreshToken" : {
          "type" : "string"
        }
      }
    },
    "Pet" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "dateBirth" : {
          "type" : "string"
        },
        "currentDisease" : {
          "type" : "string",
          "enum" : [ "RABIA", "MOQUILLO", "PULGAS", "GARRAPATAS", "LEISHMANIOSIS" ]
        }
      }
    }
  }
}